feat(eth): consolidate eip implementation into a single interface
