+++ src/interface/TWEthereum.cpp
@@ -1,13 +1,22 @@
-// Copyright © 2017-2022 Trust Wallet.
+// Copyright © 2017-2023 Trust Wallet.
 //
 // This file is part of Trust. The full Trust copyright notice, including
 // terms governing use, modification, and redistribution, is contained in the
 // file LICENSE at the root of the source code distribution tree.
 
-#include <TrustWalletCore/TWEthereumEip2645.h>
+#include "Data.h"
+#include "Ethereum/EIP1014.h"
 #include "Ethereum/EIP2645.h"
+#include <TrustWalletCore/TWEthereum.h>
 
 #include <string>
+
+TWString* TWEthereumEip1014AddressCreate2(TWString* _Nonnull fromEthAddress, TWData* _Nonnull salt, TWData* _Nonnull initCodeHash) {
+    const auto& ethAddressStr = *reinterpret_cast<const std::string*>(fromEthAddress);
+    const auto& saltData = *reinterpret_cast<const TW::Data*>(salt);
+    const auto& initCodeHashData = *reinterpret_cast<const TW::Data*>(initCodeHash);
+    return new std::string(TW::Ethereum::create2AddressString(ethAddressStr, saltData, initCodeHashData));
+}
 
 TWString* TWEthereumEip2645GetPath(TWString* ethAddress, TWString* layer, TWString* application, TWString* index) {
     const auto& ethAddressStr = *reinterpret_cast<const std::string*>(ethAddress);
