+++ tests/interface/TWHDWalletTests.cpp
@@ -17,7 +17,7 @@
 #include <TrustWalletCore/TWBase58.h>
 #include <TrustWalletCore/TWCoinType.h>
 #include <TrustWalletCore/TWSegwitAddress.h>
-#include <TrustWalletCore/TWEthereumEip2645.h>
+#include <TrustWalletCore/TWEthereum.h>
 #include <TrustWalletCore/TWEthereumMessageSigner.h>
 #include <TrustWalletCore/TWStarkExMessageSigner.h>
 #include <TrustWalletCore/TWStarkWare.h>
